Split io and meta tests into their own modules.
